Fix Qt6 compilation on Fedora rawhide.
The latest version of Qt6 has removed some overrides for QString::arg
and Qchar::Qchar causing compilation to break.  Change the existing
casts on these function calls to use a different override that still
exists.

(cherry picked from commit 80ebec6e2760f5d2a5faccdf3904cb35514cb318)

+++ mythtv/libs/libmythtv/mheg/dsmcccache.cpp
@@ -54,7 +54,7 @@ QString DSMCCCacheKey::toString() const
 {
     QString result;
     for (int i = 0; i < 4 && i < size(); i++)
-        result += QString("%1").arg(at(i), 2, 16, QChar('0'));
+        result += QString("%1").arg(static_cast<uint>(at(i)), 2, 16, QChar('0'));
     return result;
 }
